split cognitve service from main library
