+++ src/Speech/SpeechOptions.ts
@@ -1,0 +1,6 @@
+import { ISpeechRecognizer, ISpeechSynthesizer } from './SpeechInterface';
+
+export class SpeechOptions {
+    public speechRecognizer: ISpeechRecognizer;
+    public speechSynthesizer: ISpeechSynthesizer;
+}
